# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * UOU Infinite - Institutional Operating System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface RegisterBody {
  name: string;
  email: string;
  password: string;
  role: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  role: string;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface UpdateUserBody {
  name?: string;
  avatarUrl?: string;
}

export interface Course {
  id: number;
  title: string;
  code: string;
  department: string;
  /** @nullable */
  description?: string | null;
  credits: number;
  capacity: number;
  /** @nullable */
  lecturerId?: number | null;
  requiredSubjects: string;
  createdAt: string;
}

export interface CreateCourseBody {
  title: string;
  code: string;
  department: string;
  description?: string;
  credits: number;
  capacity: number;
  lecturerId?: number;
  requiredSubjects: string;
}

export interface UpdateCourseBody {
  title?: string;
  description?: string;
  capacity?: number;
  lecturerId?: number;
}

export interface Student {
  id: number;
  userId: number;
  name: string;
  email: string;
  studentId: string;
  department: string;
  level: string;
  status: string;
  /** @nullable */
  gpa?: number | null;
  /** @nullable */
  lastLoginAt?: string | null;
  loginCount: number;
  credentialToken: string;
  skills: string;
  createdAt: string;
}

export interface UpdateStudentBody {
  department?: string;
  level?: string;
  status?: string;
  skills?: string;
}

export interface StudentCredential {
  studentId: string;
  name: string;
  department: string;
  skills: string;
  credentialToken: string;
  issuedAt: string;
  qrData: string;
}

export interface AtRiskStudent {
  id: number;
  name: string;
  studentId: string;
  department: string;
  riskScore: number;
  riskReason: string;
  /** @nullable */
  lastLoginAt?: string | null;
  loginCount: number;
}

export interface Lecturer {
  id: number;
  userId: number;
  name: string;
  email: string;
  department: string;
  specialization: string;
  courseCount: number;
  createdAt: string;
}

export interface CreateLecturerBody {
  userId: number;
  department: string;
  specialization: string;
}

export interface UpdateLecturerBody {
  department?: string;
  specialization?: string;
}

export interface Admission {
  id: number;
  applicantName: string;
  email: string;
  phone: string;
  courseId: number;
  courseTitle: string;
  status: string;
  /** @nullable */
  matchPercentage?: number | null;
  /** @nullable */
  waecResults?: string | null;
  /** @nullable */
  jambScore?: number | null;
  /** @nullable */
  analysisResult?: string | null;
  submittedAt: string;
}

export interface CreateAdmissionBody {
  applicantName: string;
  email: string;
  phone: string;
  courseId: number;
  waecResults?: string;
  jambScore?: number;
}

export interface UpdateAdmissionBody {
  status?: string;
  matchPercentage?: number;
  analysisResult?: string;
}

export interface AnalyzeAdmissionBody {
  waecResults: string;
  jambScore?: number;
}

export interface SubjectScore {
  subject: string;
  grade: string;
  weight: number;
  score: number;
}

export interface AdmissionAnalysis {
  matchPercentage: number;
  subjectBreakdown: SubjectScore[];
  recommendation: string;
  eligibility: string;
}

export interface AdmissionStats {
  total: number;
  pending: number;
  approved: number;
  rejected: number;
  averageMatchPercentage: number;
}

export interface Enrollment {
  id: number;
  studentId: number;
  courseId: number;
  courseName: string;
  studentName: string;
  /** @nullable */
  grade?: string | null;
  enrolledAt: string;
}

export interface CreateEnrollmentBody {
  studentId: number;
  courseId: number;
}

export interface DashboardOverview {
  totalStudents: number;
  totalLecturers: number;
  totalCourses: number;
  totalEnrollments: number;
  pendingAdmissions: number;
  activeStudentsToday: number;
  retentionRiskCount: number;
  averageGpa: number;
}

export interface EngagementDataPoint {
  date: string;
  activeUsers: number;
  newEnrollments: number;
  submissionsCount: number;
}

export interface GeographicDataPoint {
  state: string;
  studentCount: number;
  latitude: number;
  longitude: number;
}

export interface CourseDistributionPoint {
  courseTitle: string;
  department: string;
  enrollmentCount: number;
  capacity: number;
}

export interface ActivityItem {
  id: number;
  type: string;
  message: string;
  actorName: string;
  timestamp: string;
}

export interface VerificationResult {
  valid: boolean;
  studentName?: string;
  studentId?: string;
  department?: string;
  skills?: string;
  issuedAt?: string;
  verifiedAt: string;
  blockchainHash?: string;
}

export interface SystemHealth {
  status: string;
  uptime: number;
  memoryUsage: number;
  errorCount: number;
  warningCount: number;
  aiDiagnosis: string;
  suggestions: string[];
  lastChecked: string;
}

export interface LogEntry {
  id: number;
  level: string;
  message: string;
  /** @nullable */
  context?: string | null;
  createdAt: string;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateOpenaiConversationBody {
  title: string;
}

export interface SendOpenaiMessageBody {
  content: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export type ListUsersParams = {
  role?: string;
};

export type ListStudentsParams = {
  status?: string;
  courseId?: number;
};

export type ListAdmissionsParams = {
  status?: string;
};

export type ListEnrollmentsParams = {
  studentId?: number;
  courseId?: number;
};

export type GetSystemLogsParams = {
  level?: string;
  limit?: number;
};