# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: UOU Infinite - Institutional Operating System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Users
...
  /courses:
    get:
      operationId: listCourses
      tags: [courses]
      summary: List all courses
      responses:
        "200":
          description: List of courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
    post:
      operationId: createCourse
      tags: [courses]
      summary: Create a course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCourseBody"
      responses:
        "201":
          description: Created course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /courses/{id}:
    get:
      operationId: getCourse
      tags: [courses]
      summary: Get course by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
    patch:
      operationId: updateCourse
      tags: [courses]
      summary: Update course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCourseBody"
      responses:
        "200":
          description: Updated course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
    delete:
      operationId: deleteCourse
      tags: [courses]
      summary: Delete course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Students
...
  /admissions:
    get:
      operationId: listAdmissions
      tags: [admissions]
      summary: List all admission applications
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Admission"
    post:
      operationId: createAdmission
      tags: [admissions]
      summary: Submit an admission application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAdmissionBody"
      responses:
        "201":
          description: Created application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Admission"
  /admissions/{id}:
    get:
      operationId: getAdmission
      tags: [admissions]
      summary: Get admission application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Admission application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Admission"
    patch:
      operationId: updateAdmission
      tags: [admissions]
      summary: Update admission status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAdmissionBody"
      responses:
        "200":
          description: Updated admission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Admission"
  /admissions/{id}/analyze:
    post:
      operationId: analyzeAdmission
      tags: [admissions]
      summary: AI-analyze uploaded results for course match percentage
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeAdmissionBody"
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdmissionAnalysis"
...
  /admin/logs:
    get:
      operationId: getSystemLogs
      tags: [logs]
      summary: Recent system logs
      parameters:
        - name: level
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: System logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LogEntry"

  # OpenAI Chatbot (UOU Sentinel)
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # Auth
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
      required:
        - name
        - email
        - password
        - role
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user

    # User
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - createdAt
...
    Course:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        code:
          type: string
        department:
          type: string
        description:
          type: ["string", "null"]
        credits:
          type: integer
        capacity:
          type: integer
        lecturerId:
          type: ["integer", "null"]
        requiredSubjects:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - code
        - department
        - credits
        - capacity
        - requiredSubjects
        - createdAt
    CreateCourseBody:
      type: object
      properties:
        title:
          type: string
        code:
          type: string
        department:
          type: string
        description:
          type: string
        credits:
          type: integer
        capacity:
          type: integer
        lecturerId:
          type: integer
        requiredSubjects:
          type: string
      required:
        - title
        - code
        - department
        - credits
        - capacity
        - requiredSubjects
    UpdateCourseBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        capacity:
          type: integer
        lecturerId:
          type: integer

    # Student
...
    Admission:
      type: object
      properties:
        id:
          type: integer
        applicantName:
          type: string
        email:
          type: string
        phone:
          type: string
        courseId:
          type: integer
        courseTitle:
          type: string
        status:
          type: string
        matchPercentage:
          type: ["number", "null"]
        waecResults:
          type: ["string", "null"]
        jambScore:
          type: ["integer", "null"]
        analysisResult:
          type: ["string", "null"]
        submittedAt:
          type: string
          format: date-time
      required:
        - id
        - applicantName
        - email
        - phone
        - courseId
        - courseTitle
        - status
        - submittedAt
    CreateAdmissionBody:
      type: object
      properties:
        applicantName:
          type: string
        email:
          type: string
        phone:
          type: string
        courseId:
          type: integer
        waecResults:
          type: string
        jambScore:
          type: integer
      required:
        - applicantName
        - email
        - phone
        - courseId
    UpdateAdmissionBody:
      type: object
      properties:
        status:
          type: string
        matchPercentage:
          type: number
        analysisResult:
          type: string
    AnalyzeAdmissionBody:
      type: object
      properties:
        waecResults:
          type: string
        jambScore:
          type: integer
      required:
        - waecResults
    AdmissionAnalysis:
      type: object
      properties:
        matchPercentage:
          type: number
        subjectBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/SubjectScore"
        recommendation:
          type: string
        eligibility:
          type: string
      required:
        - matchPercentage
        - subjectBreakdown
        - recommendation
        - eligibility
    SubjectScore:
      type: object
      properties:
        subject:
          type: string
        grade:
          type: string
        weight:
          type: number
        score:
          type: number
      required:
        - subject
        - grade
        - weight
        - score
...
    LogEntry:
      type: object
      properties:
        id:
          type: integer
        level:
          type: string
        message:
          type: string
        context:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - level
        - message
        - createdAt

    # OpenAI
...