# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * UOU Infinite - Institutional Operating System API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  name: string;
  email: string;
  role: string;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: Date;
}